Improve error handling, config, script setup
