+++ .example.env
@@ -1,0 +1,4 @@
+# BTC_RPC_COOKIE_PASS=long_cookie_password
+# BTC_RPC_PORT=18443
+# ELECTRS_HOST=127.0.0.1
+# ELECTRS_PORT=60401
