+++ bitcoin.conf
@@ -12,6 +12,12 @@ rpcbind=0.0.0.0
 rpcbind=127.0.0.1
 rpcport=8332
 
+[regtest]
+rpcallowip=0.0.0.0/0
+rpcbind=0.0.0.0
+rpcbind=127.0.0.1
+rpcport=18332
+
 ## Lightning options
 zmqpubrawblock=tcp://127.0.0.1:28332
 zmqpubrawtx=tcp://127.0.0.1:28333
