+++ src/rpcConstants.js
@@ -1,0 +1,93 @@
+const contentTypeHeaders = {
+  JSON: { "content-type": "application/json" },
+  PLAIN_TEXT: { "content-type": "text/plain" },
+};
+
+const errors = {
+  METHOD_NOT_SUPPORTED: {
+    code: 405,
+    httpCode: 405,
+    msg: "Method not supported",
+    statusCode: 405,
+    rpcCode: "-36200",
+    description:
+      "The server knows the request method, but the target resource doesn't support this method.",
+  },
+  CONNECTION_ERROR: {
+    code: 502,
+    httpCode: 502,
+    msg: "RPC connection error",
+    statusCode: 502,
+    description: "The server received an invalid response from the upstream server.",
+  },
+};
+
+const rpcMethods = {
+  wallet: [
+    "abandontransaction",
+    "abortrescan",
+    "addmultisigaddress",
+    "backupwallet",
+    "bumpfee",
+    "createwallet",
+    "dumpprivkey",
+    "dumpwallet",
+    "encryptwallet",
+    "getaddressesbylabel",
+    "getaddressinfo",
+    "getbalance",
+    "getbalances",
+    "getnewaddress",
+    "getrawchangeaddress",
+    "getreceivedbyaddress",
+    "getreceivedbylabel",
+    "gettransaction",
+    "getunconfirmedbalance",
+    "getwalletinfo",
+    "importaddress",
+    "importdescriptors",
+    "importmulti",
+    "importprivkey",
+    "importprunedfunds",
+    "importpubkey",
+    "importwallet",
+    "keypoolrefill",
+    "listaddressgroupings",
+    "listlabels",
+    "listlockunspent",
+    "listreceivedbyaddress",
+    "listreceivedbylabel",
+    "listsinceblock",
+    "listtransactions",
+    "listunspent",
+    "listwalletdir",
+    "listwallets",
+    "loadwallet",
+    "lockunspent",
+    "psbtbumpfee",
+    "removeprunedfunds",
+    "rescanblockchain",
+    "send",
+    "sendmany",
+    "sendtoaddress",
+    "sethdseed",
+    "setlabel",
+    "settxfee",
+    "setwalletflag",
+    "signmessage",
+    "signrawtransactionwithwallet",
+    "unloadwallet",
+    "upgradewallet",
+    "walletcreatefundedpsbt",
+    "walletlock",
+    "walletpassphrase",
+    "walletpassphrasechange",
+    "walletprocesspsbt",
+  ],
+};
+
+module.exports = {
+  contentTypeHeaders,
+  rpcMethods,
+  errors,
+};
