+++ src/server.js
@@ -3,7 +3,13 @@ const net = require("net");
 const bitcoin = require("bitcoinjs-lib");
 const BigNumber = require("bignumber.js");
 const request = require("request");
-const { rpcConfig, rpcFilters, rpcResponses, rpcHeaders } = require("./btc-rpc");
+const {
+  rpcConfig,
+  rpcFilters,
+  rpcResponses,
+  rpcHeaders,
+  rpcUtils,
+} = require("./btc-rpc");
 
 //var electrsHost = 'electrs';
 const btcCookiePass = process.env.BTC_RPC_COOKIE_PASS;
@@ -12,6 +18,7 @@ const btcRpcUrl = `127.0.0.1:${btcRpcPort}/`;
 const electrsHost = process.env.ELECTRS_HOST;
 const electrsPort = process.env.ELECTRS_PORT;
 // console.log(`INFO PROXY: btc rpc pass: ${btcCookiePass}`)
+console.log(`INFO PROXY: btc rpc url: ${btcRpcUrl}`)
 console.log(`INFO PROXY: Electrs host: ${electrsHost}:${electrsPort}`);
 
 const app = express();
@@ -81,6 +88,7 @@ const eRpc = (rpcCall, addr) => {
   });
 };
 
+// TODO: Rewrite from scratch; get tests from dependendents
 // btc rpc
 const bRpc = (rpcCall, reqOptions = {}) => {
   return new Promise((resolve, reject) => {
@@ -97,27 +105,38 @@ const bRpc = (rpcCall, reqOptions = {}) => {
       } else {
         let err;
         try {
-          err = JSON.parse(body).error;
+          if (typeof body !== "undefined") {
+            err = JSON.parse(body).error;
+          }
         } catch (e) {
+          console.error({ e, error, body });
           return reject({ code: 400, msg: "bad btc-rpc call" });
         }
         if (err != undefined) {
           return resolve(JSON.parse(body));
         } else {
+          console.error({error, err, response, body});
+          if (error.code === "ECONNREFUSED") {
+            return reject(rpcConfig.handlers.connectionError(error));
+          }
+          if (error.code === 405) {
+            return reject(rpcResponses.METHOD_NOT_SUPPORTED);
+          }
           return reject({ code: 400, msg: "bad btc-rpc call" });
         }
       }
     };
-    // Reject unallowed methods. Consider switching to whitelist for allowed methods
-    if (
-      reqOptions.useWalletFilter &&
-      rpcFilters.isWalletMethod(rpcCall.method.toLowerCase())
-    ) {
-      return reject(rpcResponses.RPC_NOT_ALLOWED);
-    }
 
     try {
-      request(options, callback);
+      // console.log({ options })
+      // Reject forbidden methods. Consider switching to whitelist for allowed methods
+      if (
+        reqOptions.useWalletFilter &&
+        rpcFilters.isWalletMethod(rpcCall?.method?.toLowerCase())
+      ) {
+        return reject(rpcResponses.METHOD_NOT_SUPPORTED);
+      }
+      return request(options, callback);
     } catch (e) {
       return reject({ code: 502 });
     }
@@ -595,17 +614,25 @@ app.get("/getblocktxs/:blockhash", (req, res) => {
     });
 });
 
-// TODO: Separate into its own file
-app.post("/btc-rpc", (req, res) => {
+// const { containsValidRpcCall, getContentType } = rpcUtils; 
+
+const handleRpc = (req, res) => {
   const reqOptions = {
     useWalletFilter: rpcConfig.flags.USE_WALLET_METHOD_FILTER,
-    headers: rpcHeaders.contentType.JSON,
+    headers: {
+      "content-type": rpcUtils.getContentType(req),
+    },
   };
-  bRpc(req.body, reqOptions)
-    .then(json => res.send(json))
-    .catch(err => {
-      res.status(err.code).end();
-    });
-});
+  if (rpcUtils.containsValidRpcCall(req)) {
+    jsonRespond(bRpc(req.body, reqOptions), identity, res);
+  } else {
+    // TODO: Add error message
+    res.status(400).end();
+  }
+};
+
+app.get("/btc-rpc", handleRpc);
+
+app.post("/btc-rpc", handleRpc);
 
 app.listen(port, () => console.log(`Electrs proxy listening on port ${port}`));
